Rename doSurvey to conductSurvey
